Isolate and Refactor multifile
This commit isolates and refactors multifile (root gets similar isolation due to needed
code-structure changes). Notably multifile has had the bulk of the complex functions
split into more contained units of work.

Errors are now bubbled up instead of calling direct os.exit() calls.
